users/serializers.py
from rest_framework import serializers
from .models import SellerProfile


class SellerProfileSerializer(serializers.ModelSerializer):
    class Meta:
        model = SellerProfile
        fields = ["avatar", "store_name", "description", "payment_info"]
