pyproject.toml
[project]
name = "marketplace"
version = "0.1.0"
description = ""
authors = [{name = "Jestuzx", email = "ycheba38@gmail.com"}]
readme = "README.md"
requires-python = ">=3.11,<4.0"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
django = "5.2.5"
django-filter = "25.2"
django-htmx = ">=1.23.2,<2.0.0"
django-render-block = ">=0.11,<0.12"
pillow = "12.0.0"
djangorestframework = "3.16.1"
alembic = "1.16.5"
amqp = "5.3.1"
anyio = "4.11.0"
asgiref = "3.9.1"
attrs = "25.4.0"
autobahn = "24.4.2"
automat = "25.4.16"
billiard = "4.2.1"
build = "1.3.0"
cachecontrol = "0.14.3"
celery = "5.5.3"
certifi = "2025.8.3"
cffi = "2.0.0"
channels = "4.3.1"
charset-normalizer = "3.4.3"
cleo = "2.1.0"
click = "8.2.1"
click-didyoumean = "0.3.1"
click-plugins = "1.1.1.2"
click-repl = "0.3.0"
colorama = "0.4.6"
constantly = "23.10.4"
crashtest = "0.4.1"
cryptography = "46.0.2"
daphne = "4.2.1"
distlib = "0.4.0"
django-split-settings = "1.3.2"
dulwich = "0.24.2"
fastjsonschema = "2.21.2"
filelock = "3.19.1"
findpython = "0.7.0"
greenlet = "3.2.4"
h11 = "0.16.0"
httpcore = "1.0.9"
httpx = "0.28.1"
hyperlink = "21.0.0"
idna = "3.10"
incremental = "24.7.2"
installer = "0.7.0"
jaraco-classes = "3.4.0"
jaraco-context = "6.0.1"
jaraco-functools = "4.3.0"
keyring = "25.6.0"
kombu = "5.5.4"
mako = "1.3.10"
markupsafe = "3.0.2"
more-itertools = "10.8.0"
msgpack = "1.1.1"
numpy = "2.3.2"
packaging = "25.0"
pbs-installer = "2025.9.18"
pkginfo = "1.12.1.2"
platformdirs = "4.4.0"
poetry = "2.2.1"
poetry-core = "2.2.1"
prompt-toolkit = "3.0.52"
psycopg2-binary = "2.9.10"
pyasn1 = "0.6.1"
pyasn1-modules = "0.4.2"
pycparser = "2.23"
pyopenssl = "25.3.0"
pyproject-hooks = "1.2.0"
python-dateutil = "2.9.0.post0"
pywin32-ctypes = "0.2.3"
rapidfuzz = "3.14.1"
requests = "2.32.5"
requests-toolbelt = "1.0.0"
service-identity = "24.2.0"
setuptools = "80.9.0"
shellingham = "1.5.4"
six = "1.17.0"
sniffio = "1.3.1"
sqlalchemy = "2.0.43"
sqlparse = "0.5.3"
tomlkit = "0.13.3"
trove-classifiers = "2025.9.11.17"
twisted = "25.5.0"
txaio = "25.9.2"
typing-extensions = "4.15.0"
tzdata = "2025.2"
urllib3 = "2.5.0"
vine = "5.1.0"
virtualenv = "20.34.0"
wcwidth = "0.2.13"
zope-interface = "8.0.1"
zstandard = "0.25.0"
social-auth-app-django = "^5.6.0"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.3.0"
isort = "^6.0.1"
pre-commit = "^4.4.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 88
